fix(demo): #138 error messages — screen-share wording, cert hint, sanitizeReason
Follow-up to #209 addressing its review findings:

- friendlyMessage had no knowledge of camera vs screen, so a denied
  screen-share read "Camera or microphone access was denied." Thread the
  MediaSourceType in (PublishBoard passes sourceType()) and branch the
  permission/not-found wording.
- friendlyConnError and the webtransport arm told the user to check
  VITE_CERT_HASH even when the cert was correctly configured (the cert
  branch returns early, so this only runs when cert is fine). Drop the
  cert mention there.
- cleanRaw didn't strip control characters (contradicting the module
  docstring) and duplicated sanitizeReason. Consolidate sanitizeReason
  into errors.ts and have cleanRaw delegate to it; ConnectionStatus and
  ScenarioView now import it from there.

Note on the SubscribeErrorCode classification: #209's review flagged the
TrackNotFound branch as dead code, but that was incorrect. The relay
resets the subscribe stream with the MoQ code
(handler.go:241 -> receive_subscribe_stream.go cancelStreamWithError,
code = SubscribeErrorCodeNotFound = 3), and @qumo/moq surfaces that as a
WebTransportStreamError whose .code is 3 — so moqCode() === TrackNotFound
matches and the "No stream at this path yet" message fires. The .code
classification is retained as-is.

Refs #138 (already closed by #209).
