+++ playground/src/errors.ts
@@ -1,10 +1,11 @@
 import { SubscribeErrorCode } from "@qumo/moq";
+import type { MediaSourceType } from "./publish/media.ts";
 
 // Friendly, actionable error messages for the demo (issue #138).
 //
 // The publish/subscribe boards used to surface bare "Error: <message>" text —
-// often an opaque DOMException string or a MoQ error code the user can't act
-// on. This module maps the common, recognizable failure cases to a short
+// often an opaque DOMException string or a MoQ internal message the user can't
+// act on. This module maps the common, recognizable failure cases to a short
 // sentence that tells the user what went wrong and what to do next. Anything
 // it can't classify falls back to a cleaned-up first line of the raw message
 // (no stack traces, no control characters).
@@ -12,23 +13,46 @@ import { SubscribeErrorCode } from "@qumo/moq";
 // Keep the messages self-contained — the UI renders them verbatim, without a
 // leading "Error:" prefix.
 
+// Defensively clean an untrusted/free-form reason before display. SolidJS
+// escapes text interpolation (so there's no HTML-injection sink), but relay
+// reasons and library messages can still contain control characters or run
+// long — strip the control chars and clamp the length so they can't flood the
+// status bar. Falls back to `fallback` when nothing remains.
+export function sanitizeReason(reason: string | undefined, fallback: string): string {
+	const isPrintable = (ch: string) => {
+		const c = ch.codePointAt(0)!;
+		return c >= 0x20 && c !== 0x7f;
+	};
+	const cleaned = Array.from(reason ?? "")
+		.filter(isPrintable)
+		.join("")
+		.trim()
+		.slice(0, 200);
+	return cleaned || fallback;
+}
+
 // Reduce a raw message to a single safe line: take the first line, drop a
-// leading "Error:" wrapper, trim, and clamp the length.
+// leading "Error:" wrapper, then strip control characters and clamp the length
+// via sanitizeReason (so the "no control characters" promise above holds).
 function cleanRaw(input: string): string {
 	const firstLine = input.split("\n")[0]?.trim() ?? "";
 	const stripped = firstLine.replace(/^error:\s*/i, "");
-	return stripped.slice(0, 200);
+	return sanitizeReason(stripped, "");
 }
 
-// MoQ subscribe/track errors carry their numeric code on `.code`. It's a
-// plain number on the wire, so read it defensively rather than isinstance.
+// MoQ stream errors surface as a WebTransportStreamError carrying the peer's
+// reset code on `.code` (the relay resets the subscribe stream with the MoQ
+// SubscribeErrorCode — e.g. TrackNotFound — rather than sending a SUBSCRIBE_ERR
+// message). It's a plain number, so read it defensively rather than instanceof.
 function moqCode(err: unknown): number | undefined {
 	const code = (err as { code?: unknown })?.code;
 	return typeof code === "number" ? code : undefined;
 }
 
-// Map a thrown error to a friendly, actionable message.
-export function friendlyMessage(err: unknown): string {
+// Map a thrown error to a friendly, actionable message. `source` (when known)
+// distinguishes a camera/microphone failure from a screen-share failure, since
+// getUserMedia and getDisplayMedia throw the same DOMException names.
+export function friendlyMessage(err: unknown, source?: MediaSourceType): string {
 	const raw = err instanceof Error ? err.message : String(err);
 	const name = (err as { name?: string } | null)?.name;
 
@@ -38,13 +62,17 @@ export function friendlyMessage(err: unknown): string {
 	switch (name) {
 		case "NotAllowedError":
 		case "SecurityError":
-			return "Camera or microphone access was denied. Allow access in your browser's site permissions, then click Start again.";
+			return source === "screen"
+				? "Screen-share access was denied. Allow it in your browser's site permissions, then click Start again."
+				: "Camera or microphone access was denied. Allow access in your browser's site permissions, then click Start again.";
 		case "NotFoundError":
-			return "No camera or microphone was found. Connect a device and try again.";
+			return source === "screen"
+				? "No screen or window was shared. Click Start and pick one to share."
+				: "No camera or microphone was found. Connect a device and try again.";
 		case "NotReadableError":
-			return "Your camera or microphone is busy. Close the other app using it, then try again.";
+			return "Your camera, microphone, or screen is busy. Close the other app using it, then try again.";
 		case "OverconstrainedError":
-			return "The selected quality isn't supported by your camera. Try a lower resolution or framerate.";
+			return "The selected quality isn't supported by your device. Try a lower resolution or framerate.";
 		case "AbortError":
 			return "Media access was cancelled. Click Start to try again.";
 		case "NotSupportedError":
@@ -55,8 +83,10 @@ export function friendlyMessage(err: unknown): string {
 			return "This operation isn't supported by your browser.";
 	}
 
-	// MoQ subscribe failures. TrackNotFound is the common one for a demo:
-	// the subscriber asked for a path nobody is publishing to.
+	// MoQ subscribe failures. The relay resets the subscribe stream with the MoQ
+	// SubscribeErrorCode, which @qumo/moq surfaces as a WebTransportStreamError
+	// whose `.code` is that code. TrackNotFound is the common demo case: the
+	// subscriber asked for a path nobody is publishing to.
 	const code = moqCode(err);
 	if (code !== undefined) {
 		if (code === SubscribeErrorCode.TrackNotFound) {
@@ -72,7 +102,7 @@ export function friendlyMessage(err: unknown): string {
 
 	// WebTransport / TLS handshake noise from the transport layer.
 	if (/webtransport|certificate|cert hash|tls|handshake|quic/i.test(raw)) {
-		return "Could not connect to the relay. Check that it's running and that VITE_CERT_HASH is set, then reload.";
+		return "Could not connect to the relay. Check that it's running and reload the page.";
 	}
 
 	const cleaned = cleanRaw(raw);
